Revert ":art: 엔티티 / 엔티티 생성 메서드 추가"
This reverts commit 470ec5a4ce78d71de358ff4d9384996b1089c28d.
