+++ src/main/java/com/findbug/findbugbackend/domain/Alarm.java
@@ -1,6 +1,6 @@
-package com.findbug.findbugbackend.domain.alarm;
+package com.findbug.findbugbackend.domain;
 
-import com.findbug.findbugbackend.domain.member.Member;
+import com.findbug.findbugbackend.domain.camera.Camera;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -11,30 +11,24 @@ import lombok.extern.jackson.Jacksonized;
 
 import java.time.LocalDateTime;
 
+
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @SuperBuilder
 @Jacksonized
 @Getter
-public class MemberAlarm {
+public class Alarm {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_alarm_id")
+    @Column(name = "alarm_id")
     private Long id;
 
-    @ManyToOne
+    private LocalDateTime detectedTime;
+
+    @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "alarm_id")
-    private Alarm alarm;
+    private Camera camera;
 
-    @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    public static MemberAlarm createMemberAlarm(Alarm alarm, Member member) {
-        return builder()
-                .alarm(alarm)
-                .member(member)
-                .build();
-    }
 }
