+++ src/main/java/com/findbug/findbugbackend/domain/AlarmBug.java
@@ -1,5 +1,7 @@
-package com.findbug.findbugbackend.domain.bug;
+package com.findbug.findbugbackend.domain;
 
+
+import com.findbug.findbugbackend.domain.bug.Bug;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -8,30 +10,29 @@ import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
 import lombok.extern.jackson.Jacksonized;
 
+import java.time.LocalDateTime;
+
+
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @SuperBuilder
 @Jacksonized
 @Getter
-public class BugInformation {
+public class AlarmBug {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "bug_information_id")
+    @Column(name = "alarm_bug_id")
     private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "alarm_id")
+    private Alarm alarm;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "bug_id")
     private Bug bug;
 
-    private String image;
-    private String description;
+    private LocalDateTime bugDetectedTime;
 
-    public static BugInformation createBugInformation(Bug bug, String image, String description) {
-        return builder()
-                .bug(bug)
-                .image(image)
-                .description(description)
-                .build();
-    }
 }
