+++ src/main/java/com/findbug/findbugbackend/domain/PestControlCompany.java
@@ -1,6 +1,7 @@
 package com.findbug.findbugbackend.domain;
 
 import com.findbug.findbugbackend.domain.member.Member;
+import com.findbug.findbugbackend.domain.valueTable.Address;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -15,10 +16,10 @@ import lombok.extern.jackson.Jacksonized;
 @SuperBuilder
 @Jacksonized
 @Getter
-public class Company {
+public class PestControlCompany {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "company_id")
+    @Column(name = "pest_control_company_id")
     private Long id;
     private String name;
     private String email;
@@ -34,5 +35,5 @@ public class Company {
 
     private String companyDescription;
     private String businessDescription;
-    private String logoImage;
+    private byte[] logoImage;
 }
