+++ src/main/java/com/findbug/findbugbackend/domain/bug/BugPreventionInfo.java
@@ -1,0 +1,4 @@
+package com.findbug.findbugbackend.domain.bug;
+
+public class BugPreventionInfo {
+}
