+++ src/main/java/com/findbug/findbugbackend/domain/bug/BugTerminationInfo.java
@@ -1,0 +1,4 @@
+package com.findbug.findbugbackend.domain.bug;
+
+public class BugTerminationInfo {
+}
