+++ src/main/java/com/findbug/findbugbackend/domain/camera/Camera.java
@@ -20,20 +20,12 @@ public class Camera {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "camera_id")
     private Long id;
-
-    @Enumerated(EnumType.STRING)
-    private CameraType type;
-
-    @Enumerated(EnumType.STRING)
-    private CameraSoftwareVersion softwareVersion;
-
     private String name;
 
-    public static Camera createCamera(CameraType type, CameraSoftwareVersion softwareVersion, String name) {
-        return builder()
-                .type(type)
-                .softwareVersion(softwareVersion)
-                .name(name)
-                .build();
-    }
+    @Enumerated(EnumType.STRING)
+    private CameraType cameraType;
+
+    @Enumerated(EnumType.STRING)
+    private CameraSoftwareVersion cameraSoftwareVersion;
+
 }
