+++ src/main/java/com/findbug/findbugbackend/domain/member/LoginMethod.java
@@ -1,0 +1,5 @@
+package com.findbug.findbugbackend.domain.member;
+
+public enum LoginMethod {
+    KAKAO, GOOGLE, NAVER
+}
