+++ src/main/java/com/findbug/findbugbackend/domain/valueTable/Address.java
@@ -1,4 +1,4 @@
-package com.findbug.findbugbackend.domain;
+package com.findbug.findbugbackend.domain.valueTable;
 
 
 import jakarta.persistence.Embeddable;
